# backend/docs/openapi.yaml
openapi: 3.0.3
info:
  title: SportsMatch API
  version: 1.0.0

# 1. Environment Configuration
servers:
  - url: http://localhost:3000/api/v1
    description: local environment
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

  # 3. Reusable Data Models (Schemas)
  schemas:
...
    VenueClaimRequest:
      type: object
      required: [contact_person, contact_title, contact_phone, contact_email]
      properties:
        contact_name:
          type: string
          description: Optional override. If omitted, backend uses authenticated user's display name.
        contact_person: { type: string }
        contact_title: { type: string }
        contact_phone: { type: string }
        contact_email: { type: string }
        note: { type: string }


# 4. API Endpoints
paths:
...
  /venues:
    get:
      tags: [Venues]
      summary: List venues
      parameters:
        - in: query
          name: lat
          schema: { type: number }
        - in: query
          name: lng
          schema: { type: number }
        - in: query
          name: radiusKm
          schema: { type: number }
      responses:
        "200":
          description: ok
  /venues/{venueId}:
    get:
      tags: [Venues]
      summary: Get venue by id
      parameters:
        - in: path
          name: venueId
          required: true
          schema: { type: string, format: uuid }
      responses:
        "200":
          description: ok
  /venues/{venueId}/claim:
    post:
      tags: [Venues]
      summary: Request to claim a venue (Lead generation)
      security: [{ BearerAuth: [] }]
      parameters:
        - in: path
          name: venueId
          required: true
          schema: { type: string, format: uuid }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/VenueClaimRequest'
      responses:
        "200":
          description: ok
  /admin/venues:
    get:
      tags: [Admin]
      summary: List all venues for governance
      security: [{ BearerAuth: [] }]
      parameters:
        - in: query
          name: search
          schema: { type: string }
        - in: query
          name: limit
          schema: { type: integer, default: 50 }
        - in: query
          name: offset
          schema: { type: integer, default: 0 }
      responses:
        "200":
          description: ok
  /admin/venue-claims:
    get:
      tags: [Admin]
      summary: List venue claims for admin review
      security: [{ BearerAuth: [] }]
      parameters:
        - in: query
          name: status
          description: Filter claims by status. Omit for all statuses.
          schema:
            type: string
            enum: [pending, approved, rejected]
        - in: query
          name: limit
          schema: { type: integer, default: 50 }
        - in: query
          name: offset
          schema: { type: integer, default: 0 }
      responses:
        "200":
          description: ok
  /admin/venue-claims/{claimId}/approve:
    post:
      tags: [Admin]
      summary: Approve claim and assign official email
      security: [{ BearerAuth: [] }]
      parameters:
        - in: path
          name: claimId
          required: true
          schema: { type: string, format: uuid }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [officialEmail]
              properties:
                officialEmail: { type: string, format: email }
      responses:
        "200":
          description: ok
  /admin/venue-claims/{claimId}/reject:
    post:
      tags: [Admin]
      summary: Reject a pending claim
      security: [{ BearerAuth: [] }]
      parameters:
        - in: path
          name: claimId
          required: true
          schema: { type: string, format: uuid }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [reason]
              properties:
                reason: { type: string, minLength: 5 }
      responses:
        "200":
          description: ok
  /admin/venue-claims/{claimId}/revoke:
    post:
      tags: [Admin]
      summary: Revoke an approved claim
      security: [{ BearerAuth: [] }]
      parameters:
        - in: path
          name: claimId
          required: true
          schema: { type: string, format: uuid }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [reason]
              properties:
                reason: { type: string, minLength: 5 }
      responses:
        "200":
          description: ok
...
  /admin/venues/{venueId}:
    get:
      tags: [Venue Portal]
      summary: Get venue profile for owner/admin portal
      security: [{ BearerAuth: [] }]
      parameters:
        - in: path
          name: venueId
          required: true
          schema: { type: string, format: uuid }
      responses:
        "200":
          description: ok
    put:
      tags: [Venue Portal]
      summary: Update venue profile for owner/admin portal
      security: [{ BearerAuth: [] }]
      parameters:
        - in: path
          name: venueId
          required: true
          schema: { type: string, format: uuid }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [logo_url, opening_hours, facilities, playing, services, supply]
              properties:
                logo_url: { type: string }
                description: { type: string }
                social_links:
                  type: object
                  additionalProperties: { type: string }
                opening_hours:
                  type: array
                  minItems: 7
                  maxItems: 7
                  items:
                    type: object
                    required: [is_open, open_at, close_at]
                    properties:
                      is_open: { type: boolean }
                      open_at: { type: string, nullable: true, example: "06:00" }
                      close_at: { type: string, nullable: true, example: "22:00" }
                facilities:
                  type: object
                  properties:
                    parking: { type: boolean }
                    restroom: { type: boolean }
                    shower: { type: boolean }
                    changing_room: { type: boolean }
                    seating_area: { type: boolean }
                playing:
                  type: object
                  properties:
                    lighting: { type: boolean }
                    indoor: { type: boolean }
                    outdoor: { type: boolean }
                services:
                  type: object
                  properties:
                    equipment_rental: { type: boolean }
                    coaching: { type: boolean }
                supply:
                  type: object
                  properties:
                    water: { type: boolean }
                    vending_machine: { type: boolean }
                    contactless_pay: { type: boolean }
                    wifi: { type: boolean }
                spaces:
                  type: array
                  items:
                    type: object
                    required: [name, supported_sports]
                    properties:
                      name: { type: string }
                      supported_sports:
                        type: array
                        items: { type: string }
      responses:
        "200":
          description: ok

    patch:
      tags: [Admin]
      summary: Patch venue display and operator fields
      security: [{ BearerAuth: [] }]
      parameters:
        - in: path
          name: venueId
          required: true
          schema: { type: string, format: uuid }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name_display: { type: string }
                address_display: { type: string }
                operator_name: { type: string }
                operator_email: { type: string, format: email }
                operator_role: { type: string }
                operator_phone: { type: string }
      responses:
        "200":
          description: ok
  /admin/venues/{venueId}/suspend:
    post:
      tags: [Admin]
      summary: Suspend a venue from operator management
      security: [{ BearerAuth: [] }]
      parameters:
        - in: path
          name: venueId
          required: true
          schema: { type: string, format: uuid }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [reason]
              properties:
                reason: { type: string, minLength: 5 }
      responses:
        "200":
          description: ok
  /admin/venues/{venueId}/unsuspend:
    post:
      tags: [Admin]
      summary: Restore a suspended venue to claimed status
      security: [{ BearerAuth: [] }]
      parameters:
        - in: path
          name: venueId
          required: true
          schema: { type: string, format: uuid }
      responses:
        "200":
          description: ok
...